+++ src/lib/Circle.ts
@@ -4,10 +4,12 @@ import {Body} from './Body.js';
  * A circle used to detect collisions
  */
 export class Circle extends Body {
+	override readonly _polygon = false as const;
+	override readonly _circle = true as const;
+	override readonly _point = false as const;
+
 	radius: number;
 	scale: number;
-
-	override _circle = true;
 
 	/**
 	 * x: The starting X coordinate
@@ -21,18 +23,5 @@ export class Circle extends Body {
 
 		this.radius = radius;
 		this.scale = scale;
-	}
-
-	/**
-	 * Draws the circle to a CanvasRenderingContext2D's current path
-	 * 		context: The context to add the arc to
-	 */
-	draw(context: CanvasRenderingContext2D): void {
-		const x = this.x;
-		const y = this.y;
-		const radius = this.radius * this.scale;
-
-		context.moveTo(x + radius, y);
-		context.arc(x, y, radius, 0, Math.PI * 2);
 	}
 }